requirements.txt
python-socketio
appdirs